+++ index.js
@@ -16,17 +16,34 @@ local.env.setDispatchWrapper(function(request, origin, dispatch) {
 		local.http.reqheader(request, 'link', { href:'httpl://storage.env/'+request.urld.host, rel:'http://grimwire.com/rel/storage' });
 		// ^ when multiple peers' servers enter the namespace, this will direct the Worker to the correct user's storage
 
-		// attach cookies
-		var storedCookies = storageServer.getItem(request.urld.host, '.cookies');
-		var reqCookies = {}; // extract cookie values
-		if (storedCookies && storedCookies.items) {
-			for (var k in storedCookies.items) {
-				reqCookies[k] = storedCookies.items[k].value || storedCookies.items[k];
+		var reqCookies = {};
+		// attach client cookies
+		if (origin instanceof local.client.Region) {
+			var clientCookies = origin.cookies[request.urld.authority];
+			if (clientCookies) {
+				for (var k in clientCookies) {
+					reqCookies[k] = clientCookies[k].value || clientCookies[k];
+					// ^ cookies may be given as a single value or as an object with {value:...}
+
+					// add flagged values to the query object
+					if (clientCookies[k].query)
+						request.query[k] = (typeof request.query[k] == 'undefined') ? clientCookies[k].value : request.query[k];
+				}
+			}
+		}
+		// attach session cookies
+		var sessionCookies = storageServer.getItem(request.urld.host, '.cookies');
+		if (sessionCookies && sessionCookies.items) {
+			for (var k in sessionCookies.items) {
+				if (k in reqCookies)
+					continue;
+
+				reqCookies[k] = sessionCookies.items[k].value || sessionCookies.items[k];
 				// ^ cookies may be given as a single value or as an object with {value:...}
 
 				// add flagged values to the query object
-				if (storedCookies.items[k].query)
-					request.query[k] = request.query[k] || storedCookies.items[k].value;
+				if (sessionCookies.items[k].query)
+					request.query[k] = (typeof request.query[k] == 'undefined') ? sessionCookies.items[k].value : request.query[k];
 			}
 		}
 		local.http.reqheader(request, 'cookie', reqCookies);
@@ -49,6 +66,9 @@ local.env.setDispatchWrapper(function(request, origin, dispatch) {
 			if (!storedCookies.items || typeof storedCookies.items != 'object')
 				storedCookies.items = {}; // save us from corruption
 			for (var k in cookies) {
+				if (cookies[k].scope && cookies[k].scope != 'session')
+					continue;
+
 				if (cookies[k] === null)
 					delete storedCookies.items[k];
 				else